third_party/blink/web_tests/fast/filesystem/directory-entry-to-uri-expected.txt

Obtaining URL from DirectoryEntry.

On success, you will see a series of "PASS" messages, followed by "TEST COMPLETE".

PASS testDirectoryURL is 'filesystem:file:///temporary/testDirectory'
PASS successfullyParsed is true

TEST COMPLETE
